% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzorsaActivity.R
\docType{data}
\name{AzorsaActivity}
\alias{AzorsaActivity}
\title{Kinase Activity on Phosphorylating Tau Protein}
\format{
A data frame with 3243 rows and 6 variables:
\describe{
\item{Kinase}{The name of the Kianse}
\item{siRNA}{weight of the diamond, in carats}
\item{Run}{weight of the diamond, in carats}
\item{AvgCells}{weight of the diamond, in carats}
\item{phospho_tau}{weight of the diamond, in carats}
\item{total_tau}{weight of the diamond, in carats}
}
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1186\%2F1471-2164-11-25/MediaObjects/12864_2008_2619_MOESM1_ESM.XLS}
}
\usage{
data(AzorsaActivity)
}
\description{
Kinase Activity on Phosphorylating Tau Protein
}
\keyword{datasets}
